SIBO-468: Add regression test driver for testing example notebooks.
The driver generates a separate test each for executing and checking flake8 on the Python equivalent of any notebooks found in the examples directory.

Output from failures related to execution tests, which I had to force in order to test, is written to stdout so can be checked in the log.

The results of flake8 tests are summarized and appended to the generated Python file, which is retained if there are any errors.  Every notebook currently fails the flake8 checks.  Of concern, though, is some notebooks (e.g., getting_started.ipynb) won't be able to pass due to generated get_ipython calls for ipython magics.