SIBO-468: Added missing (tox) dependencies and temporarily disabled flake8 test generation.
Tox initialization had to be modified to include notebook-related dependencies and disable sphinx apidoc of notebook tests (at least until additional dependencies can be addressed).

Modified the jupyter notebook test generation so that the generated flake8 tests will skip "E501 line too long" (due to magic cels) and, temporarily, disable generation of the flake8 tests (see SIBO-481).

Finally, have the bash script that runs the tests, which include flake8 on Sina, to skip the examples directory so as not to preclude the generation and execution of flake8 tests on a per-file basis.  (We should actually consider whether we really want flake8 failures to preclude running actual tests.  In other words, shouldn't we be running the actual tests before worrying about PEP8 style checks?)