SIBO-467: Added get_example_path utility and added to the notebooks
Introduced the SINA_TEST_KERNEL environment variable to facilitate Jupyter Notebook testing.  The environment variable is used to provide a name for the test kernel that is built when the tests are run as well as to determine if a test database path is returned from the new sina get_example_path utility.  The utility assumes a naming convention (i.e., appending a suffix such as '-new') is used to when building a test database for example notebooks.

This work includes the introduction of a set_jupyter_kernel.sh script to actually build the kernel for Tox.

Finally, I switched the order of the clean-* entries for the Makefile clean target.