SIBO-598: Make build and testing of Jupyter optional.
Ran across a post that mentioned tox defaults to setup.py under some circumstances, which is not resolved using 'pip install' as some other posts suggest due to our wheelhouse constraints.

So it occurred to me this weekend to try to use the Makefile to build since it does all the associated "heavy lifting".  That seems to have done the trick, at least for me on quartz.